requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai
requests
pydantic
streamlit>=1.24.0
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
plotly==5.18.0
xgboost==2.0.3
numpy==1.26.2
geopy==2.4.1
pytest==8.0.0
